Rework the way startup code is executed.
The issue was that an indented text in there would raise an
IndentationError. I tried a bunch of different things with the code
module to get around this. None of them worked, but manipulating the
namespace with an execution of the startup code before the console is
started and then passing the namespace as the locals to the new console
works.